MRP: fix dropped LeaveAll bit in VectorHeader serialization
VectorHeader.serialize computed the LeaveAll bit as `leaveAllEvent.rawValue <<
13`, where rawValue is a UInt8. Shifting a UInt8 left by 13 overshoots its
8-bit width, so Swift's smart shift yields 0 and the LeaveAll bit (0x2000) was
silently dropped from every transmitted vector.

Two consequences on the wire:
 - A LeaveAll carrying values lost its flag entirely, so peers were never told
   to re-register — the LeaveAll/re-registration mechanism (10.7.6) was broken.
 - A LeaveAll-only vector (NumberOfValues=0) serialized to a 0x0000 VectorHeader,
   byte-identical to the attribute-list EndMark. A strict MRPDU parser then reads
   the attribute list as empty.
   (pyshark: "No such field vector_attribute" on Listener/Talker Failed messages).

Widen to UInt16 before shifting so the bit lands at position 13. The existing
testVectorHeaderSerialization had frozen the buggy output ([0x02,0x34] decoding
back to NullLeaveAllEvent); corrected to the proper 0x2234 encoding.

+++ Sources/MRP/Base/PDU.swift
@@ -118,7 +118,9 @@ struct VectorHeader: Sendable, Equatable, Hashable, SerDes {
   }
 
   func serialize(into serializationContext: inout SerializationContext) throws {
-    let value = UInt16(leaveAllEvent.rawValue << 13) | UInt16(numberOfValues & 0x1FFF)
+    // widen before the shift: leaveAllEvent.rawValue is UInt8, so `rawValue << 13` in UInt8
+    // arithmetic overshifts its 8-bit width and drops the LeaveAll bit (10.8.2.6)
+    let value = (UInt16(leaveAllEvent.rawValue) << 13) | (numberOfValues & 0x1FFF)
     serializationContext.serialize(uint16: value)
   }
 }
